Relax levenshtein check for falky tests...
